Fix no template id that match valid share
There are several things that can result in the above issue:
1. TP send too many new templates and we process a newer template before
   and older have been processed but we receive the share for the older
   that will never have a mathcing template id since will be discarded
2. upstream take more then 10 second to authorize a job but we find a
   valid share right after the we see the template.

This commmit add a lock so that we can not process newer templates until
old one have ben declared and acceted. It also make sure that we are not
going to process 2 templates that are sent within a very short amount of
time (we only get the newer).

It increase the maximum amount of time to get the template id from 10 to
15 sec.

It do not return an Irrecoverable error if we are not able to get the
template id within 15 sec from when the share have been received. But it
only log that we have a stale share.