Fix firmware compatibility
Use a randomized job id for every notify that we send downstream also
when the notify is for an already sent job (this happens when we do
difficulty adj). It seems that some firmware when receive 2 ore more
time a notify with the same job id it start producing invalid shares and
eventually it close the connection.